fix(test): add missing extra_deny_subjects arg to evaluate() calls
engine.evaluate() signature changed to require a second argument
(Option<Vec<Subject>>), but 12 call sites in integration tests were
not updated, causing E0061 compilation errors on master.

Also split permission integration tests into a separate file to
satisfy the 500-line file limit enforced by the pre-commit hook.

Source: CI
Type: fix
